In the middle of merging the functions/entities/calls from the ViewController (UI Layer) for all API/Persistence operations down to the Business Layer using the EntityManager service class.
This class manages all of the API calls and the Persistence Layer calls, which include populating/updating the database and also retrieving the stored data for in-memory use to populate the UI elements.

Commit also includes some WIP changes to the Entities themselves. Clarifications will need to be made regarding which entities are specifically used by the API to match the JSON Encoder and which of those fields are relevant/necessary for storage in the Persistence Layer.

Lastly, the RevatureAPI class calls have been updated to include a more expansive error message description. The requestLocation method also has proper usage of the header and parameter fields for the AlamoFire API Library.