Finalized the first section of the EntityManager service class. This service class handles making requests to the API, storing the results, parsing them into entities, passing them to the persistence layer, and providing access to them for the UI elements.
Updated the Entity strut for APIRoomCall regarding the Campus reference. Was initially going to be a different kind of Campus, but will make the campus object returned from the API the new defacto standard because all of that information is necessary when submitting that information to the backend server later.

Additionally, updating some comments to better reflect changes that need to be made and tasks that need to be completed.