Updated App Delegate to contain global single instance of the EntityManager service class. Added comments for where to execute certain statements/functionality.
Changed activeUser to class private with getter (for later verification and error checking). Changed reference to getter method in Login Controller.

Added all the basic functionality for the programmatically generated Trainer workflow table view/cells. Need to re-order some data for more clarification of location/campus'.

Updated EntityClasses to better reflect recently changed API calls. Data in the classes needs to reflect the match for the new APIs so the entities had to be updated.

Added In-memory getters for all the information from the Service class. These methods will always be used for calling data from the application to display (whether or not this data is filled from API or SQLite is determined earlier).