Fixed the EntityManager email reference.
Added new prototype ViewController, RevatureTrainerDetailsVC.
Added TableView didSelectRowAt functionality for the TrainerWorkflow selection page. This selection redirects to the prototype class that is instantiated at runtime containing all the details of the individual trainer that is selected from the table.

Added functions to the login controller to hide the keyboard when password entry is complete (blocks the Login button on 6s Plus iPhone models).

Refactored the constraints (again) for the 6s Plus using its size class.

(Always) updated project file with references to new classes.